Move handlers to its own go file
